Rename Flipper::Adapters::Poll::Poller to Flipper::Poller
