+++ lib/flipper.rb
@@ -155,6 +155,7 @@ require 'flipper/instrumenters/noop'
 require 'flipper/identifier'
 require 'flipper/middleware/memoizer'
 require 'flipper/middleware/setup_env'
+require 'flipper/poller'
 require 'flipper/registry'
 require 'flipper/type'
 require 'flipper/types/actor'
